Stable surge hook V2 deployment prep.
