+++ v3/deprecated/20250121-v3-stable-surge/output/base.json
@@ -1,0 +1,5 @@
+{
+  "StableSurgePoolFactory": "0x4fb47126Fa83A8734991E41B942Ac29A3266C968",
+  "StableSurgeHook": "0xb2007B8B7E0260042517f635CFd8E6dD2Dd7f007",
+  "MockStablePool": "0xf551bD21bc0E568F2a0f54F9213E3dA7336e5625"
+}
