+++ v3/deprecated/20250121-v3-stable-surge/output/mainnet.json
@@ -1,0 +1,5 @@
+{
+  "StableSurgePoolFactory": "0xD53F5d8d926fb2a0f7Be614B16e649B8aC102D83",
+  "StableSurgeHook": "0xb18fA0cb5DE8cecB8899AAE6e38b1B7ed77885dA",
+  "MockStablePool": "0xc05ff9851c2a000F3C319D2986D8712317583B79"
+}
