--- v3/tasks/20250121-v3-stable-surge/output/gnosis.json
@@ -1,5 +1,0 @@
-{
-  "StableSurgePoolFactory": "0x268E2EE1413D768b6e2dc3F5a4ddc9Ae03d9AF42",
-  "StableSurgeHook": "0xe4f1878eC9710846E2B529C1b5037F8bA94583b1",
-  "MockStablePool": "0x825DC69A2FC808620eEd7377B19D204e6944c675"
-}
