--- v3/tasks/20250121-v3-stable-surge/output/sepolia.json
@@ -1,5 +1,0 @@
-{
-  "StableSurgePoolFactory": "0xD516c344413B4282dF1E4082EAE6B1081F3b1932",
-  "StableSurgeHook": "0x30CE53fA38a1399F0CA158b5c38362c80E423bA9",
-  "MockStablePool": "0xD8D88e186AA7cfa3929dF79ab5a5f02316e7d77B"
-}
