+++ v3/tasks/20250403-v3-stable-surge-hook-v2/index.ts
@@ -1,0 +1,14 @@
+import { Task, TaskRunOptions } from '@src';
+import { StableSurgeHookDeployment } from './input';
+
+export default async (task: Task, { force, from }: TaskRunOptions = {}): Promise<void> => {
+  const input = task.input() as StableSurgeHookDeployment;
+
+  const args = [
+    input.Vault,
+    input.DefaultMaxSurgeFeePercentage,
+    input.DefaultSurgeThresholdPercentage,
+    input.HookVersion,
+  ];
+  await task.deployAndVerify('StableSurgeHook', args, from, force);
+};
