+++ v3/tasks/20250403-v3-stable-surge-hook-v2/readme.md
@@ -1,0 +1,8 @@
+# 2025-04-03 - V3 Stable Surge Hook (V2)
+
+Deployment of the `StableSurgeHook` as a standalone contract, which can be plugged in to pools coming from any pool factory for flexibility.
+
+## Useful Files
+
+- [Code](https://github.com/balancer/balancer-v3-monorepo/commit/).
+- [`StableSurgeHook` artifact](./artifact/StableSurgeHook.json)
